chore: renamed files to improve organisation
